banner cosmetics and welcome ux define
